Fix bugs

+++ emulator/ppu/gba/opengl/tiles.c
@@ -7,7 +7,7 @@
 void OpenGlTilesReload(OpenGlTiles* context, const GbaPpuMemory* memory,
                        const GbaPpuRegisters* registers,
                        GbaPpuDirtyBits* dirty_bits) {
-  for (uint8_t i = 0u; i < GBA_TILE_MODE_NUM_BACKGROUND_TILE_BLOCKS; i++) {
+  for (uint32_t i = 0u; i < GBA_TILE_MODE_NUM_BACKGROUND_TILE_BLOCKS; i++) {
     if (!dirty_bits->vram.tile_mode.tiles[i] || registers->dispcnt.mode >= 3) {
       continue;
     }
@@ -46,14 +46,14 @@ void OpenGlTilesReload(OpenGlTiles* context, const GbaPpuMemory* memory,
   }
 
   uint8_t starting_tile = (registers->dispcnt.mode < 3) ? 0u : 1u;
-  for (uint8_t i = starting_tile; i < 2; i++) {
+  for (uint32_t i = starting_tile; i < 2; i++) {
     if (!dirty_bits->vram.obj_tiles[i]) {
       continue;
     }
 
     uint32_t insert_index = 0u;
-    uint16_t s_tile_base = GBA_BITMAP_MODE_NUM_OBJECT_S_TILES * i;
-    uint16_t d_tile_base = GBA_BITMAP_MODE_NUM_OBJECT_D_TILES * i;
+    uint32_t s_tile_base = GBA_BITMAP_MODE_NUM_OBJECT_S_TILES * i;
+    uint32_t d_tile_base = GBA_BITMAP_MODE_NUM_OBJECT_D_TILES * i;
     for (uint16_t t = 0u; t < GBA_BITMAP_MODE_NUM_OBJECT_S_TILES; t++) {
       for (uint8_t y = 0u; y < GBA_TILE_1D_SIZE; y++) {
         for (uint8_t x = 0u; x < GBA_TILE_1D_SIZE / 2u; x++) {
@@ -77,9 +77,9 @@ void OpenGlTilesReload(OpenGlTiles* context, const GbaPpuMemory* memory,
     glBindTexture(GL_TEXTURE_2D, context->obj_d_tiles);
     glTexSubImage2D(
         GL_TEXTURE_2D, /*level=*/0, /*xoffset=*/0,
-        /*yoffset=*/GBA_TILE_1D_SIZE * GBA_BITMAP_MODE_NUM_OBJECT_D_TILES * i,
+        /*yoffset=*/GBA_TILE_1D_SIZE * d_tile_base,
         /*width=*/GBA_TILE_1D_SIZE,
-        /*height=*/GBA_TILE_1D_SIZE * GBA_BITMAP_MODE_NUM_OBJECT_S_TILES,
+        /*height=*/GBA_TILE_1D_SIZE * GBA_BITMAP_MODE_NUM_OBJECT_D_TILES,
         /*format=*/GL_LUMINANCE, /*type=*/GL_UNSIGNED_BYTE,
         /*pixels=*/&memory->vram.mode_012.obj.d_tiles[d_tile_base]);
     glBindTexture(GL_TEXTURE_2D, 0);
@@ -165,8 +165,7 @@ void OpenGlTilesReloadContext(OpenGlTiles* context) {
   glTexImage2D(GL_TEXTURE_2D, /*level=*/0, /*internal_format=*/GL_LUMINANCE,
                /*width=*/GBA_TILE_1D_SIZE,
                /*height=*/GBA_TILE_1D_SIZE * GBA_TILE_MODE_NUM_OBJECT_D_TILES,
-               /*border=*/0, /*format=*/GL_LUMINANCE,
-               /*type=*/GL_UNSIGNED_BYTE,
+               /*border=*/0, /*format=*/GL_LUMINANCE, /*type=*/GL_UNSIGNED_BYTE,
                /*pixels=*/NULL);
 
   glBindTexture(GL_TEXTURE_2D, 0);
